pyproject.toml
[tool.poetry]
name = "vidhash"
version = "0.1.5"
description = "A package for hashing videos and checking for similarity"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/SpangleLabs/vidhash"

[tool.poetry.dependencies]
python = "^3.8"
ImageHash = "^4.2.1"
Pillow = "^9.1.1"
ffmpy3 = "^0.2.4"
numpy = "^1.22.4"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-black = "^0.3.3"
mypy = "^0.950"
black = "^22.3.0"
flake8-type-checking = "^1.5.0"
types-Pillow = "^9.0.15"
isort = "^5.10.1"
requests = "^2.27.1"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
types-requests = "^2.27.29"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
mypy_path = "vidhash/stubs"
files = "vidhash,test"

[tool.isort]
line_length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
